requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
pydantic[email]>=2.10.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
google-genai>=1.0.0

